Move try_guess_topics_name to utils
